+++ cg_tpe2/src/main/java/ar/edu/itba/cg/tpe2/core/light/PointLight.java
@@ -1,15 +1,15 @@
 package ar.edu.itba.cg.tpe2.core.light;
 
-import javax.vecmath.Point3d;
+import javax.vecmath.Point3f;
 
 import ar.edu.itba.cg.tpe2.core.colors.Specular;
 
 public class PointLight extends Light {
 
 	private Specular aSpec;
-	private Point3d p;
+	private Point3f p;
 	
-	public PointLight(String name, Specular aSpec, double power, Point3d p) {
+	public PointLight(String name, Specular aSpec, float power, Point3f p) {
 		super(ELightType.point, name,power);
 		this.aSpec = aSpec;
 		this.p = p;
@@ -19,7 +19,7 @@ public class PointLight extends Light {
 		return aSpec;
 	}
 
-	public Point3d getP() {
+	public Point3f getP() {
 		return p;
 	}
 
@@ -30,7 +30,7 @@ public class PointLight extends Light {
 
 	
 	
-	public float getFallOff(double distanceToLight) {
+	public float getFallOff(float distanceToLight) {
 //		(1./(     log(xs.+1).+1    )    ).*100;
 		if ( distanceToLight < 0 )
 			return 0;
